Move types to shared and closer to features
