+++ client/src/typings/declarations/custom.d.ts
