--- client/src/typings/group.typings.ts
@@ -1,9 +1,0 @@
-import { SelectedGame } from 'shared/typings/models/user';
-
-export interface GroupMember {
-  enteredGames: readonly SelectedGame[];
-  groupCode: string;
-  serial: string;
-  signedGames: readonly SelectedGame[];
-  username: string;
-}
