+++ client/src/utils/dragAndDrop.ts
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { moveArrayItem, insertByIndex } from 'client/utils/array';
-import { DnDUpdatedPositions, DnDMove } from 'client/typings/game.typings';
+import { DnDUpdatedPositions, DnDMove } from 'client/views/signup/signupTypes';
 import { Game } from 'shared/typings/models/game';
 
 export const reorder = (
