+++ client/src/utils/utils.typings.ts
@@ -1,0 +1,4 @@
+export interface WeekdayAndTime {
+  time: string;
+  capitalize: boolean;
+}
