+++ client/src/views/group/components/GroupMembersList.tsx
@@ -1,7 +1,7 @@
 import React, { FC, ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
-import { GroupMember } from 'client/typings/group.typings';
+import { GroupMember } from 'shared/typings/api/groups';
 
 export interface Props {
   groupMembers: readonly GroupMember[];
